+++ tests/integration/integration_test.go
@@ -233,6 +233,10 @@ func TestValidateResource(t *testing.T) {
 				"  'state' must not have the value 5.",
 			},
 		},
+		// Test scenario 10: no violations.
+		{
+			WantErrors: nil,
+		},
 	})
 }
 
@@ -286,6 +290,28 @@ func TestValidateStack(t *testing.T) {
 				"  mandatory: Prohibits setting state to 3 on dynamic resources.",
 				"  'state' must not have the value 3.",
 			},
+		},
+		// Test scenario 6: violates the fourth policy.
+		{
+			WantErrors: []string{
+				"  mandatory: Prohibits creating a RandomUuid without any 'keepers'.",
+				"  RandomUuid must not have an empty 'keepers'.",
+			},
+		},
+		// Test scenario 7: violates the fifth policy.
+		{
+			WantErrors: []string{
+				"  mandatory: Prohibits RandomString resources.",
+				"  RandomString resources are not allowed.",
+			},
+		},
+		// Test scenario 8: no violations.
+		{
+			WantErrors: nil,
+		},
+		// Test scenario 9: no violations.
+		{
+			WantErrors: nil,
 		},
 	})
 }