+++ tests/integration/validate_stack/policy-pack/index.ts
@@ -1,6 +1,8 @@
 // Copyright 2016-2019, Pulumi Corporation.  All rights reserved.
 
-import { PolicyPack } from "@pulumi/policy";
+import { PolicyPack, validateStackResourcesOfType } from "@pulumi/policy";
+
+import * as random from "@pulumi/random";
 
 new PolicyPack("validate-stack-test-policy", {
     policies: [
@@ -29,7 +31,7 @@ new PolicyPack("validate-stack-test-policy", {
 
                     if (r.type === "pulumi-nodejs:dynamic:Resource") {
                         if (r.props.state === 1) {
-                            reportViolation("'state' must not have the value 1.")
+                            reportViolation("'state' must not have the value 1.");
                         }
                     }
                 }
@@ -50,7 +52,7 @@ new PolicyPack("validate-stack-test-policy", {
 
                     if (r.type === "pulumi-nodejs:dynamic:Resource") {
                         if (r.props.state === 2) {
-                            reportViolation("'state' must not have the value 2.")
+                            reportViolation("'state' must not have the value 2.");
                         }
                     }
                 }
@@ -76,6 +78,72 @@ new PolicyPack("validate-stack-test-policy", {
                     }
                 }
             },
+        },
+        // Strongly-typed.
+        {
+            name: "randomuuid-no-keepers",
+            description: "Prohibits creating a RandomUuid without any 'keepers'.",
+            enforcementLevel: "mandatory",
+            validateStack: validateStackResourcesOfType(random.RandomUuid, (resources, args, reportViolation) => {
+                for (const r of resources) {
+                    if (!r.keepers || Object.keys(r.keepers).length === 0) {
+                        reportViolation("RandomUuid must not have an empty 'keepers'.");
+                    }
+                }
+            }),
+        },
+        // Manual strongly-typed.
+        {
+            name: "no-randomstrings",
+            description: "Prohibits RandomString resources.",
+            enforcementLevel: "mandatory",
+            validateStack: (args, reportViolation) => {
+                const resources = args.resources
+                    .map(r => r.isType(random.RandomString))
+                    .filter(r => r);
+                if (resources.length > 0) {
+                    reportViolation("RandomString resources are not allowed.");
+                }
+            },
+        },
+        // Validate other type checks work as expected.
+        {
+            name: "test-type-checks",
+            description: "Policy used to test type checks.",
+            enforcementLevel: "mandatory",
+            validateStack: (args, reportViolation) => {
+                for (const r of args.resources) {
+                    if (r.type !== "random:index/randomPassword:RandomPassword") {
+                        continue;
+                    }
+                    if (!r.isType(random.RandomPassword)) {
+                        throw new Error("`isType` did not return the expected value.");
+                    }
+                    const randomPassword = r.asType(random.RandomPassword);
+                    if (!randomPassword) {
+                        throw new Error("`asType` did not return the expected value.");
+                    }
+                    if (randomPassword.length !== 42) {
+                        throw new Error("`randomPassword.length` did not return the expected value.");
+                    }
+                }
+            },
+        },
+        // Validate that `args.resources` is filtered and matches the resources in `resources`.
+        {
+            name: "test-args-filtering",
+            description: "Policy used to test that `args.resources` is filtered and matches the resources in `resources`.",
+            enforcementLevel: "mandatory",
+            validateStack: validateStackResourcesOfType(random.RandomInteger, (resources, args, reportViolation) => {
+                if (resources.length !== args.resources.length) {
+                    throw new Error("`args.resources.length` does not match `resources.length`.");
+                }
+                for (let i = 0; i < resources.length; i++) {
+                    if (resources[i].id !== args.resources[i].props.id) {
+                        throw new Error("`resources[i].id` does not match `args.resources[i].props.id`.");
+                    }
+                }
+            }),
         },
     ],
 });