+++ tests/integration/validate_stack/program/index.ts
@@ -1,6 +1,7 @@
 // Copyright 2016-2019, Pulumi Corporation.  All rights reserved.
 
 import * as pulumi from "@pulumi/pulumi";
+import * as random from "@pulumi/random";
 import { Resource } from "./resource";
 
 const config = new pulumi.Config();
@@ -29,5 +30,32 @@ switch (testScenario) {
     case 5:
         // Violates the third policy.
         const c = new Resource("c", { state: 3 });
+        break;
+
+    case 6:
+        // Violates the fourth policy.
+        const r1 = new random.RandomUuid("r1");
+        break;
+
+    case 7:
+        // Violates the fifth policy.
+        const r2 = new random.RandomString("r2", {
+            length: 10,
+        });
+        break;
+
+    case 8:
+        // Create a resource to test the strongly-typed helpers.
+        const r3 = new random.RandomPassword("r3", {
+            length: 42,
+        });
+        break;
+
+    case 9:
+        // Create several resources of the same type to validate
+        // resource filtering by type.
+        const x = new random.RandomInteger("x", { min: 0, max: 10 });
+        const y = new random.RandomInteger("y", { min: 0, max: 10 });
+        const z = new random.RandomInteger("z", { min: 0, max: 10 });
         break;
 }